+++ NEWS.md
@@ -1,4 +1,5 @@
 # MIPr (development version)
 
+* Add functions to convert amino acid abbreviations (@arisp99, #4).
 * Increase efficiency of reading in files (@arisp99, #2).
 * Added a `NEWS.md` file to track changes to the package.